requirements.txt
asn1crypto==0.24.0
cachetools==2.1.0
cffi==1.11.5
cryptography==2.3.1
fuzzywuzzy==0.17.0
google-api-python-client==1.7.3
google-auth==1.5.0
google-auth-httplib2==0.0.3
httplib2==0.11.3
idna==2.7
oauth2client==4.1.2
pyasn1==0.4.3
pyasn1-modules==0.2.1
pycparser==2.18
PyYAML==3.13
rsa==3.4.2
six==1.11.0
sortedcontainers==2.1.0
uritemplate==3.0.0
